refactor: PlayerRepository player domain 으로 정리
