+++ modules/bukkit-database/src/main/kotlin/kr/hqservice/framework/database/repository/player/lock/impl/SwitchDefermentLock.kt
@@ -1,11 +1,11 @@
-package kr.hqservice.framework.database.lock.impl
+package kr.hqservice.framework.database.repository.player.lock.impl
 
 import kotlinx.coroutines.*
 import kr.hqservice.framework.bukkit.core.HQBukkitPlugin
 import kr.hqservice.framework.bukkit.core.coroutine.element.UUIDCoroutineContextElement
 import kr.hqservice.framework.bukkit.core.coroutine.component.coroutinescope.HQCoroutineScope
 import kr.hqservice.framework.bukkit.core.coroutine.extension.coroutineContext
-import kr.hqservice.framework.database.lock.DefermentLock
+import kr.hqservice.framework.database.repository.player.lock.DefermentLock
 import kr.hqservice.framework.global.core.component.Component
 import kr.hqservice.framework.global.core.component.Qualifier
 import kr.hqservice.framework.global.core.component.Singleton
