+++ modules/bukkit-database/src/main/kotlin/kr/hqservice/framework/database/repository/player/registry/PlayerRepositoryRegistry.kt
@@ -1,6 +1,6 @@
-package kr.hqservice.framework.database.registry
+package kr.hqservice.framework.database.repository.player.registry
 
-import kr.hqservice.framework.database.repository.PlayerRepository
+import kr.hqservice.framework.database.repository.player.PlayerRepository
 import kr.hqservice.framework.global.core.component.HQSimpleComponent
 
 interface PlayerRepositoryRegistry : HQSimpleComponent {
